requirements.txt
# Requirements.txt
This file contains the dependencies required for this app to run properly. 

numpy==1.24.3
requests==2.31.0
spacy==3.7.2
streamlit==1.30.0
transformer==4.40.1
torch==2.2.2
